+++ lib/models/questions/animal_question.dart
@@ -1,0 +1,11 @@
+class AnimalQuestion {
+  final String imageAsset;
+  final List<String> options;
+  final String correctAnswer;
+
+  AnimalQuestion({
+    required this.imageAsset,
+    required this.options,
+    required this.correctAnswer,
+  });
+}
