+++ lib/models/questions/bird_question.dart
@@ -1,0 +1,11 @@
+class BirdQuestion {
+  final String imageAsset;
+  final List<String> options;
+  final String correctAnswer;
+
+  BirdQuestion({
+    required this.imageAsset,
+    required this.options,
+    required this.correctAnswer,
+  });
+}
