+++ lib/models/questions/flower_question.dart
@@ -1,0 +1,11 @@
+class FlowerQuestion {
+  final String imageAsset;
+  final List<String> options;
+  final String correctAnswer;
+
+  FlowerQuestion({
+    required this.imageAsset,
+    required this.options,
+    required this.correctAnswer,
+  });
+}
