+++ lib/models/questions/occupation_question.dart
@@ -1,0 +1,11 @@
+class OccupationQuestion {
+  final String imageAsset;
+  final List<String> options;
+  final String correctAnswer;
+
+  OccupationQuestion({
+    required this.imageAsset,
+    required this.options,
+    required this.correctAnswer,
+  });
+}
