+++ lib/models/questions/part_question.dart
@@ -1,0 +1,11 @@
+class PartQuestion {
+  final String imageAsset;
+  final List<String> options;
+  final String correctAnswer;
+
+  PartQuestion({
+    required this.imageAsset,
+    required this.options,
+    required this.correctAnswer,
+  });
+}
