+++ lib/models/questions/planet_question.dart
@@ -1,0 +1,11 @@
+class PlanetQuestion {
+  final String imageAsset;
+  final List<String> options;
+  final String correctAnswer;
+
+  PlanetQuestion({
+    required this.imageAsset,
+    required this.options,
+    required this.correctAnswer,
+  });
+}
