+++ lib/models/questions/question.dart
@@ -1,0 +1,13 @@
+class Questions {
+  String question;
+  List<String> options;
+  String? image;
+  int answer;
+
+  Questions({
+    required this.question,
+    required this.options,
+    required this.answer,
+    this.image,
+  });
+}
